Fix compiler warnings
-Wall -Wextra -Wno-stringop-overflow -Wno-unused-parameter -Wnon-virtual-dtor -pedantic

+++ src/overpass_api/core/type_tags.h
@@ -351,7 +351,7 @@ struct Tag_Index_Global_Has_Key_Functor {
   inline bool operator()(const void* data) const
    {
      char* k = ((int8*)data + 4);
-     int len = unalignedLoad<uint16>(data);
+     uint16 len = unalignedLoad<uint16>(data);
      return (len == key.length() && std::strncmp(k, key.c_str(), len) == 0);
    }
 
@@ -368,10 +368,10 @@ struct Tag_Index_Global_Has_Value_Functor {
   inline bool operator()(const void* data) const
    {
      // char* k = ((int8*)data + 4);
-     int key_len = unalignedLoad<uint16>(data);
+    uint16 key_len = unalignedLoad<uint16>(data);
 
      char* v = ((int8*)data + 4 + key_len);
-     int value_len = unalignedLoad<uint16>((uint16*)data + 1);
+     uint16 value_len = unalignedLoad<uint16>((uint16*)data + 1);
 
      return (value_len == value.length() && std::strncmp(v, value.c_str(), value_len) == 0);
    }
